+++ pdata/pmetric/generated_metrics_test.go
@@ -1523,9 +1523,9 @@ func TestSummaryDataPoint_Sum(t *testing.T) {
 
 func TestSummaryDataPoint_QuantileValues(t *testing.T) {
 	ms := NewSummaryDataPoint()
-	assert.Equal(t, NewValueAtQuantileSlice(), ms.QuantileValues())
-	internal.FillTestValueAtQuantileSlice(internal.ValueAtQuantileSlice(ms.QuantileValues()))
-	assert.Equal(t, ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice()), ms.QuantileValues())
+	assert.Equal(t, NewSummaryDataPointValueAtQuantileSlice(), ms.QuantileValues())
+	internal.FillTestSummaryDataPointValueAtQuantileSlice(internal.SummaryDataPointValueAtQuantileSlice(ms.QuantileValues()))
+	assert.Equal(t, SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice()), ms.QuantileValues())
 }
 
 func TestSummaryDataPoint_Flags(t *testing.T) {
@@ -1536,41 +1536,41 @@ func TestSummaryDataPoint_Flags(t *testing.T) {
 	assert.Equal(t, testValFlags, ms.Flags())
 }
 
-func TestValueAtQuantileSlice(t *testing.T) {
-	es := NewValueAtQuantileSlice()
+func TestSummaryDataPointValueAtQuantileSlice(t *testing.T) {
+	es := NewSummaryDataPointValueAtQuantileSlice()
 	assert.Equal(t, 0, es.Len())
-	es = newValueAtQuantileSlice(&[]*otlpmetrics.SummaryDataPoint_ValueAtQuantile{})
+	es = newSummaryDataPointValueAtQuantileSlice(&[]*otlpmetrics.SummaryDataPoint_ValueAtQuantile{})
 	assert.Equal(t, 0, es.Len())
 
 	es.EnsureCapacity(7)
-	emptyVal := newValueAtQuantile(&otlpmetrics.SummaryDataPoint_ValueAtQuantile{})
-	testVal := ValueAtQuantile(internal.GenerateTestValueAtQuantile())
+	emptyVal := newSummaryDataPointValueAtQuantile(&otlpmetrics.SummaryDataPoint_ValueAtQuantile{})
+	testVal := SummaryDataPointValueAtQuantile(internal.GenerateTestSummaryDataPointValueAtQuantile())
 	assert.Equal(t, 7, cap(*es.getOrig()))
 	for i := 0; i < es.Len(); i++ {
 		el := es.AppendEmpty()
 		assert.Equal(t, emptyVal, el)
-		internal.FillTestValueAtQuantile(internal.ValueAtQuantile(el))
+		internal.FillTestSummaryDataPointValueAtQuantile(internal.SummaryDataPointValueAtQuantile(el))
 		assert.Equal(t, testVal, el)
 	}
 }
 
-func TestValueAtQuantileSlice_CopyTo(t *testing.T) {
-	dest := NewValueAtQuantileSlice()
+func TestSummaryDataPointValueAtQuantileSlice_CopyTo(t *testing.T) {
+	dest := NewSummaryDataPointValueAtQuantileSlice()
 	// Test CopyTo to empty
-	NewValueAtQuantileSlice().CopyTo(dest)
-	assert.Equal(t, NewValueAtQuantileSlice(), dest)
+	NewSummaryDataPointValueAtQuantileSlice().CopyTo(dest)
+	assert.Equal(t, NewSummaryDataPointValueAtQuantileSlice(), dest)
 
 	// Test CopyTo larger slice
-	ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice()).CopyTo(dest)
-	assert.Equal(t, ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice()), dest)
+	SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice()).CopyTo(dest)
+	assert.Equal(t, SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice()), dest)
 
 	// Test CopyTo same size slice
-	ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice()).CopyTo(dest)
-	assert.Equal(t, ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice()), dest)
+	SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice()).CopyTo(dest)
+	assert.Equal(t, SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice()), dest)
 }
 
-func TestValueAtQuantileSlice_EnsureCapacity(t *testing.T) {
-	es := ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice())
+func TestSummaryDataPointValueAtQuantileSlice_EnsureCapacity(t *testing.T) {
+	es := SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice())
 	// Test ensure smaller capacity.
 	const ensureSmallLen = 4
 	expectedEs := make(map[*otlpmetrics.SummaryDataPoint_ValueAtQuantile]bool)
@@ -1603,24 +1603,24 @@ func TestValueAtQuantileSlice_EnsureCapacity(t *testing.T) {
 	assert.Equal(t, expectedEs, foundEs)
 }
 
-func TestValueAtQuantileSlice_MoveAndAppendTo(t *testing.T) {
+func TestSummaryDataPointValueAtQuantileSlice_MoveAndAppendTo(t *testing.T) {
 	// Test MoveAndAppendTo to empty
-	expectedSlice := ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice())
-	dest := NewValueAtQuantileSlice()
-	src := ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice())
+	expectedSlice := SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice())
+	dest := NewSummaryDataPointValueAtQuantileSlice()
+	src := SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice())
 	src.MoveAndAppendTo(dest)
-	assert.Equal(t, ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice()), dest)
+	assert.Equal(t, SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice()), dest)
 	assert.Equal(t, 0, src.Len())
 	assert.Equal(t, expectedSlice.Len(), dest.Len())
 
 	// Test MoveAndAppendTo empty slice
 	src.MoveAndAppendTo(dest)
-	assert.Equal(t, ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice()), dest)
+	assert.Equal(t, SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice()), dest)
 	assert.Equal(t, 0, src.Len())
 	assert.Equal(t, expectedSlice.Len(), dest.Len())
 
 	// Test MoveAndAppendTo not empty slice
-	ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice()).MoveAndAppendTo(dest)
+	SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice()).MoveAndAppendTo(dest)
 	assert.Equal(t, 2*expectedSlice.Len(), dest.Len())
 	for i := 0; i < expectedSlice.Len(); i++ {
 		assert.Equal(t, expectedSlice.At(i), dest.At(i))
@@ -1628,51 +1628,51 @@ func TestValueAtQuantileSlice_MoveAndAppendTo(t *testing.T) {
 	}
 }
 
-func TestValueAtQuantileSlice_RemoveIf(t *testing.T) {
+func TestSummaryDataPointValueAtQuantileSlice_RemoveIf(t *testing.T) {
 	// Test RemoveIf on empty slice
-	emptySlice := NewValueAtQuantileSlice()
-	emptySlice.RemoveIf(func(el ValueAtQuantile) bool {
+	emptySlice := NewSummaryDataPointValueAtQuantileSlice()
+	emptySlice.RemoveIf(func(el SummaryDataPointValueAtQuantile) bool {
 		t.Fail()
 		return false
 	})
 
 	// Test RemoveIf
-	filtered := ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice())
+	filtered := SummaryDataPointValueAtQuantileSlice(internal.GenerateTestSummaryDataPointValueAtQuantileSlice())
 	pos := 0
-	filtered.RemoveIf(func(el ValueAtQuantile) bool {
+	filtered.RemoveIf(func(el SummaryDataPointValueAtQuantile) bool {
 		pos++
 		return pos%3 == 0
 	})
 	assert.Equal(t, 5, filtered.Len())
 }
 
-func TestValueAtQuantile_MoveTo(t *testing.T) {
-	ms := ValueAtQuantile(internal.GenerateTestValueAtQuantile())
-	dest := NewValueAtQuantile()
+func TestSummaryDataPointValueAtQuantile_MoveTo(t *testing.T) {
+	ms := SummaryDataPointValueAtQuantile(internal.GenerateTestSummaryDataPointValueAtQuantile())
+	dest := NewSummaryDataPointValueAtQuantile()
 	ms.MoveTo(dest)
-	assert.Equal(t, NewValueAtQuantile(), ms)
-	assert.Equal(t, ValueAtQuantile(internal.GenerateTestValueAtQuantile()), dest)
+	assert.Equal(t, NewSummaryDataPointValueAtQuantile(), ms)
+	assert.Equal(t, SummaryDataPointValueAtQuantile(internal.GenerateTestSummaryDataPointValueAtQuantile()), dest)
 }
 
-func TestValueAtQuantile_CopyTo(t *testing.T) {
-	ms := NewValueAtQuantile()
-	orig := NewValueAtQuantile()
+func TestSummaryDataPointValueAtQuantile_CopyTo(t *testing.T) {
+	ms := NewSummaryDataPointValueAtQuantile()
+	orig := NewSummaryDataPointValueAtQuantile()
 	orig.CopyTo(ms)
 	assert.Equal(t, orig, ms)
-	orig = ValueAtQuantile(internal.GenerateTestValueAtQuantile())
+	orig = SummaryDataPointValueAtQuantile(internal.GenerateTestSummaryDataPointValueAtQuantile())
 	orig.CopyTo(ms)
 	assert.Equal(t, orig, ms)
 }
 
-func TestValueAtQuantile_Quantile(t *testing.T) {
-	ms := NewValueAtQuantile()
+func TestSummaryDataPointValueAtQuantile_Quantile(t *testing.T) {
+	ms := NewSummaryDataPointValueAtQuantile()
 	assert.Equal(t, float64(0.0), ms.Quantile())
 	ms.SetQuantile(float64(17.13))
 	assert.Equal(t, float64(17.13), ms.Quantile())
 }
 
-func TestValueAtQuantile_Value(t *testing.T) {
-	ms := NewValueAtQuantile()
+func TestSummaryDataPointValueAtQuantile_Value(t *testing.T) {
+	ms := NewSummaryDataPointValueAtQuantile()
 	assert.Equal(t, float64(0.0), ms.Value())
 	ms.SetValue(float64(17.13))
 	assert.Equal(t, float64(17.13), ms.Value())